XBee_Sensor: Centralize number of sensors property
Read the setting once, in the XBee base init method, instead of tracking
it in the physical version and fetching it from the settings in various
loops in the simulator class.

Also make a read-only public property for the setting, so that it can be
reused by others to know how large the network is supposed to be.