Control_Panel_Devices_View: Move update from xbee thread
Qt does not really like updating the UI from other threads than its
event loop or its timers. The devices discovery callback is called from
the xbee thread at least for the physical version. This can cause
a warning about the unsafe use of pixmaps on other threads.

To amend this, add a timer that is started by the refresh method which
checks whether an update flag is set. If so, update the tree view with
new data. The vehicle refresh callback then only needs to set this flag.

We stop the timer when all the devices have joined, since that is the
only way they can update at this point. Also stop the timer when
switching views in between.

Change the xbee_sensor_physical runner so that it does not immediately
start measuring. This allows us to actually use the discovery without
getting snowed in.