Control_Panel: Start caching data when switching
Add methods to the view for receiving cached data, and providing data to
be cached. These are dictionaries containing member variable data that
should be small enough to not cause a huge memory footprint, but
important enough to keep around.

For the devices view, use this to keep the list of devices after
switching. When switching back, the devices are immediately/still
visible, instead of having to wait for discovery. This data might also
be used later on by other views to know which vehicles are available, so
that they can disable inputs, etc.

Other use cases may be to keep the current reconstruction data or to
keep the waypoints table populated when switching.