Waypoints view: Cache data, use number of sensors
Convert waypoints view to a list and stacked widget, like the settings
view. Use import and export to implement the save and load of the view
when switching views. Rewrite the export make it easier to read all the
table data, and treat cells with an empty item widget the same as a cell
with no data at all. Handle string/integer indices correctly. Add a nice
error message when no waypoints at all are found.