XBee_Packet: Pack strings and other objects
Add a specification for the settings packet communications, which is to
be used by the Settings view save sender. The specification relies on
specially defined formats that allow us to pack a variable-length string
into the XBee packet. Additionally, we can send arbitrary JSON-encodable
objects that are compressed by Zlib. In the arbitrary object format, for
some basic types (bool, int, float and string), we use the smaller
packed format instead of the slower and bulkier compression steps.

This allows us to send settings in the packets without immediate concern
about packet sizes. Even the current value for "servo_pins", a list of
dictionaries, fits in the packet at less than 80 bytes in total.