Send settings from control panel to vehicles
Hook up the XBee sender to the confirmation dialog of the settings view.
If we want to send them to vehicles, start the sender that works in
a similar way to the waypoints sender. After it is done sending, save
them to the ground station settings (if we want to) as well.

Hook up the settings receiver to the Environment so that all missions
accept settings during the custom startup (non-measurement) phase.
Update the test to just use the created receiver.

Clean up the XBee_Packet serialization/unserialization error handling,
so that it becomes more obvious which specification and which field is
responsible for causing an error. Also makes the code a bit more
readable by moving blocks of code into their own method.