Reconstruction view: use only packets with valid locations
Packets with invalid locations are shown in the table and the graph (for
the sake of overview), but are not used anymore in the actual
reconstruction.

Additionally, rename `_label` to `_image_label` for clarity and decrease
the number of singular values to compute because of the change in this
patch.