Improve reset behavior
Bind the start/stop buttons sooner so that they are more responsive
instead of having a lot of initialization time where we cannot reset the
vehicle in case of problems such as infinite loops. Also, we do not want
to destroy the threads on the infrared sensor thread, since that does
not stop the main thread correctly in this case. Instead, interrupt the
main thread, which then actually disables all the other threads.

Also enable line buffering for the service script output so that we see
the output during the execution.