Settings: Test whether defaults.json is correct
This unit test loads the registration information for all components and
settings from the defaults JSON file. It then checks the existence of
certain keys in each component and setting field, which can depend on
the type of the setting. A helpful type checker object wraps some logic
and nice output formatting for the type-specific checks.

This test is helpful when creating new settings, since it ensures that
the fields are correct for all purposes (command line arguments and
control panel) and do not contain, e.g., fields meant for another type.

We do not test actual values of the defaults or related fields. The test
can be stress-tested itself by changing some types of settings that also
have fields meant for specific types, or removing a required field.