Mission: XBee synchronization at waypoints in AUTO mode
Initial work to make the vehicles wait for each other at their waypoint
until measurements are valid. When enabled, this adds "wait" commands in
between waypoints for AUTO missions. The vehicle autopilot then stops
automatic traversal of the waypoints indefinitely. One can only continue
the mission by updating the waypoint, which happens when the measurement
is valid.

For MAVLink vehicles, the "wait" command is implemented using a loiter
command. For the robot vehicles, we use markers that make the command an
invalid waypoint and have special behavior. Due to this, the vehicle
remains at the wait waypoint indefinitely, and the location remains
valid while waiting.