Mission: Add wait commands in XBee mission, clean up waypoints
Make a base Mission_Auto method for adding waypoints to the vehicle's
mission commands. This method now handles all the conversions needed for
locations in specific geometries, and adds the wait command if enabled.

Use this method for adding waypoints in Mission_XBee, so that the wait
commands are added when XBee waypoint synchronization is enabled. Also
track the received waypoint count ourselves instead of calculating the
next index in a fragile and vehicle-dependent manner.