Tests: Clean up Environment-based tests
Create an EnvironmentTestCase which handles setting up the Environment
object with given arguments, mocking the infrared sensor if enabled, and
handling all the test case inheritances.

This simplifies the setup and import requirements in various tests, and
makes it easier to make more mission tests that need an Environment.

Move the DTR mock from the Mission_Cycle test to USBManagerTestCase and
generalize it to DTR and RTS state for all serial devices, since the
normal methods do not function when we mock the devices as virtual PTYs.
This ensures we do not have any problems with the USB serial devices in
any other USB-based tests as well.