Mission_XBee: Add tests
Create a test that ensures correct behavior from the XBee mission.
Included are tests for the setup method and for the packet callbacks of
clearing waypoints, adding a waypoint and done completing the waypoints.
The latter also tests whether XBee waypoint synchronization works
correctly, i.e., the vehicle waits at a waypoint with a wait command
until the location packet is valid.

The test uncovered a problem in an edge case that was introduced in
fe3551d5332d12e0f41a42210720332d370c6294: when clearing a non-empty
mission, the next index tracking the waypoint count was not reset.