XBee measurements: Wait for all sensors or certain of them
Revamp the valid location side effect in the Environment: We now only
consider a measurement to be valid if all required vehicle sensors have
a valid location, instead of just the current vehicle and another one.
This is mostly important when we have more than two vehicles, which may
have different measurement synchronization schemes. We can also specify
a specific set of sensors to synchronize with; a new method handles this
as well as invalidating the measurements, ensuring we do not consider
a measurement to be valid too quickly.

To make this possible, the XBee's valid callback needs the sensor ID of
the other vehicle to be passed. This requires changing all the callbacks
that did not yet have it. Also clean up setting the sensor ID of the
rssi_broadcast in the base class rather than the specializations, since
there is no need to do it this way anymore.

Add tests for the measurement validity, and ensure that the tests have
a correct number of sensors for their behavioral checks, and the XBee ID
is a vehicle ID instead of zero.