Waypoints view: Add more columns to tables
Rename the two existing column headers to the more descriptive and more
correct "north" and "east" (which should be good enough in all geometry
settings) and add two new columns for "altitude" and "wait for vehicle".
These two column are correctly imported and exported and do not require
to be filled (by having defaults that are not None). They are not yet
used in the XBee sender code though.

Also fix the code that was supposed to remove all the rows from the
table which did not remove all since it was changing the row indexes in
between. Reverse the removal order to remove all rows. Make it a method
of the table widget. It is odd that no such method exists in Qt core.