Waypoints view: Change default columns tab order
Instead of a normal ordering by-row and by-column, skip the columns that
have defaults defined for them since they are not immediately required.
Only when we cannot proceed any further, we cycle through the columns
with defaults vertically.

This means that we preserve the fast way of inputting the north/east
columns like before, and adding altitudes and synchronization data can
be added afterward in one go.