Send new XBee waypoint data and handle in mission
Correctly cast the waypoint columns since the coordinates may be
floating point, but the wait ID is integer. While we are at it, clean up
the error messages from exporting.

Add altitude and wait ID to the specification and all (test) packets.
When receiving it in Mission_XBee, add them to Mission_Auto.add_waypoint
so that we can check for measurement validity of the required sensors
during the wait command.