Reconstruction view: simplify drawing the image
Make use of Matplotlib's `FigureCanvas` which we can directly insert
into the layout so we do not have to work with a separate `QImage`
anymore.

Matplotlib determines a good image size by itself, so we do not need to
resize it anymore and therefore the viewer dimensions settings is no
longer needed.