Planning view: Integrate positions display
Add a list widget and a stacked layout for selecting different solution
individuals with planned sensor positions. Each item in the list widget
has a placeholder text and another label, which is either text info for
the objectives or feasibility, or it is a miniature plot image. This is
only updated for the Pareto front plot, the currently shown item or all
the position plots when the algorithm is done. Ensure that plots are of
correct size and are deleted when they are replaced. The maximum number
of open plots in matplotlib is adjusted to accommodate all plots. Alter
the dimensions of the plot and the progress bar to fit nicely, and make
the points in the Pareto front pickable so that the sensor positions of
that solution are shown. Add lines between list items for clear layout.