Reconstruction: move option checking to base class and create more tests
Each buffer requires options to be specified, so we can move the check
for that to the base class so we do not have to duplicate it.

Furthermore this patch creates more unit tests for the buffers. We
extend the existing buffer unit tests with tests for the presence of
options and tests for the correctness of the properties. We create new
unit tests for the dataset buffer, dump buffer and stream buffer.