Reconstruction view: replace the toolbar with a sidebar
The sidebar consists of multiple panels (tabs), one for each data
source. Each panel contains default items (provided by the base class)
and items specific for each data source (provided by the subclass).

Every item in the panel is registered and provides a lambda function
for obtaining the value of the widget in a structured manner.