XBee sensor simulator/physical: be consistent with RSSI sign
The datasets and dumps used negative RSSI values, which is the correct
way to handle them. However, the XBee sensors reported positive RSSI
values. To be consistent, make the XBee sensors report negative RSSI
values instead. This is correct because the XBee specification PDF
mentions that the RSSI value is reported in -dBm, so we need to change
the sign ourselves to get the value in dBm.