Reconstruction planning: Snap point inside network during planning
Add a parameter to Snap_To_Boundary and the Weight_Matrix that creates
it to enable them to handle points that are placed inside the network.
Normally, such sensor points are invalid and must be discarded, but the
discrete planning problem is not smart enough to know that it cannot
place points there. Therefore, enabling this state makes the snapper
place these points away from each other, taking slope and closest
boundaries into account. This even works if one point is inside the
network and the other is outside or on the boundary, regardless of which
one is start or end point.

This makes it possible to run the discrete planning problem with a much
lower number of measurements and unsnappable rate, since we lose less
points to these unsnappable problems. We do need to round the points to
grid coordinates in this case, but this should overall improve the
solutions made by the planning problem.

The tests make sure that all (edge) cases are handled correctly. Some
other new accessors of the Problem, Weight_Matrix and Snap_To_Boundary
consolidate code and reduces some small duplication. Finally, Clean up
imports and tuple/enum definitions for Snap_To_Boundary and its tests.