Reconstruction dataset/dump buffer: generate XBee packets on demand
Previously we generated all XBee packets at once when starting the
reconstruction process, but as the Utah datasets are quite large, this
made the UI freeze for approximately 10 seconds.

Instead of generating all XBee packet objects at once, we now only
generate them on demand, i.e., when the `get` method is called. This
makes the reconstruction for datasets and dumps start immediately.