Reconstruction viewer: implement stop button
The start button is converted to a toggle button that allows us to not
only start the reconstruction, but also stop it (instead of having to
close the entire application like before).

This work means that we must now clear the image when we restart. We
also fix a bug in the graph's `clear` method that caused old curves to
not be removed.

Furthermore, we lower the pause time to make reconstruction faster.