Algorithm: Fix missing dominated layers
The `sort_nondominated` method as an argument that makes the function
optimize for speed by not calculating all the layers, but just the first
nondominated one. This is used by the algorithm to find dominated or
noncontributing solutions more quickly. However, it expected that the
other layers were filled, thus never removing dominated solutions.

This had a serious impact on the quality since the nondominated
solutions would be discarded before the dominated ones, thus decreasing
the possibility of moving the Pareto front toward minimization.

Do what the algorithm means to do and put the dominated solutions all in
one layer without further ado so that the dominated solutions are
actually being replaced.