Geometry, Mission: Documentation, race condition
Clean up the documentation in Geometry in preparation for creating
a special Grid version of the geometry. This eventually spread out
toward the Mission classes, which like Geometry had comments which
spread out for too long over one line and had outdated information.

Unrelated to this is a fix specifically for the Mission_Cycle test:
there could be a race condition between the vehicle's AUTO mode and
check_waypoint when updating the current waypoint index. Since this
appears only in tests, hopefully this is an adequate and sane fix.