Reconstruction dataset buffer: do not use 0-based indexing
The Utah datasets use node IDs that are 0-based. Our sensor IDs are
1-based because we reserved ID 0 for the ground station. Therefore, for
each packet we get from the dataset buffer, we need to increment the
destination sensors ID with 1 to get the expected vehicle ID in the
table and to make the plot work properly.

Furthermore, remove the first and last (incomplete) sweeps from the
dataset. This is also recommended in the README file delivered with the
Utah datasets.