Reconstruction viewer: implement chunked rendering
We no longer render an image after each (valid) measurement for
performance reasons. Instead, we render an image after each chunk of
(valid) measurements. The size of a chunk is a setting that can be tuned
per data source. The default value of this setting is based on the
dataset source.

This patch therefore increases packet throughput while keeping the
refresh rate for the image relatively high.