Reconstruction viewer: reduce the effect of timers
Since chunked rendering is now implemented, the effect of the timers has
been reduced dramatically. We wait for a chunk to be complete before
rendering a new image, which takes some time to happen. The timers that
remain are therefore only used to stop the loop from taking up 100% CPU,
just like in other places in the codebase.