Reconstruction coordinator: implement RSSI replacement policy
This patch makes the weight matrix and RSSI vector minimal by replacing
RSSI values for existing links instead of constantly adding them.
Another advantage of this approach is that old data is automatically
being replaced with more recent data, making the code more suitable for
moving entities in the network.