Planning view: Fix memory leaks, clean up drawing
The core problem of some memory leaks in the planning view was that the
plots and axes were closed, but the FigureCanvas of each plot item was
not cleaned up when replacing it in a following run. This meant that the
plots and all their data remained in memory, but were not referenced
from anywhere. We now free this using a Qt object cleanup handler and
ensure we close all components of the plots, list items and the graph,
including axes, figures, canvases, plot references and graph plot data.

Split out cleanup, initialization and setup parts to their own methods
and call the correct ones when initializing, (re)starting and switching
views, which now also stops the planning runner thread. During setup, do
not draw the Pareto front immediately because that causes a weird flash
of a misplaced window.

Disable the XBee during the planning run and enable it afterward so that
it does not interfere with the thread, which should increase performance
somewhat. This was mostly useful during debugging since the XBee sensors
create some stray traceback/exception objects that are collected during
a different cycle.