Tests: Improve thread tests
Ensure that any created threads are destroyed and there are no other
latent threads. Simplify the thread manager lookup in ThreadableTestCase
objects by cleaning up the EnvironmentTestCase. Fix tests that did not
call the parent setUp when they should. The distance sensor physical
test setup is simplified by moving the module mocking and configuration
to the EnvironmentTestCase of which it is now a subclass.