XBee sensor: Handle halted physical sensor correctly
When we deactivate, we only set the serial connection to None since the
USB manager should close the connection. This however means that the
XBee sensor library can still write to it when sending messages. Instead
do the same as in the simulated version and raise an error when we are
deactivated.