Settings widget: Handle validation errors better
When an input has an unacceptable value, then it is silently ignored in
the settings and planning views. Since we consider it to be the same
value as it was before, we cannot detect this externally. Therefore,
make SettingsWidget.get_values also create a dictionary of key-values
showing which values are allowed or not. We use this in the settings
view to display them as warnings in the save dialog, and as a critical
error in the planning view, stopping the runner before it even started.