Settings: Convert reconstruction view to use table widget
Replace the panels with conversion functions with the settings table
widget with validation and built-in conversions. Use settings instead of
options in the dependent objects, namely the Buffers and Stream_Recorder
(which now uses the controller for XBee rather than the central widget).
The settings component and buffer related to each tab is now stored
centrally, allowing for easier initialization. The source components
have the reconstruction component as their parent, thus the options for
the canvas drawing (color map, interpolation) and other settings are now
also available. The Stream_Buffer needs to register the XBee as well,
which is now wrapped in a convenience function that detects whether we
want to record. Make the origin and size in the Buffers tuples rather
than mutable lists.

Add a method to the settings widget to get all values, so that setting
them back to default actually propagates. This method also simplifies
the allowed value check.

For file formats, allow overriding the decision on using full or short
values so that all reconstruction dataset/dump/stream files have their
full file name.

Update the buffer test so they use settings and pass with these changes.