Reconstruction view: Show reconstructor settings
Add a stacked widget that conditionally shows settings based on the
combo box selection for the reconstructor, and update settings based on
the settings provided. The only reconstructor with any settings is the
truncated SVD one, but other reconstructors may have settings later on
and it is useful to quickly change the number of singular values from
within the control panel.