Loading view: Switch to simulated XBee
Make it possible to start the control panel as usual without arguments,
but change the XBee type while the loading view is showing the progress.
This makes it possible to start simulation when one does not have
a physical XBee around but forgot to use the argument trick.

Also add documentation in the Controller.