Makefile: Add some useful commands
Add a command to start the control panel, which is now the default when
running 'make' instead of 'clean'. 'make test' is now a short way to run
'python2 test.py'. Also add commands for working with the Raspberry Pi
service. Depending on the command and the OS, these commands need
superuser rights, i.e., use 'sudo make ...'.