Import design for laser cutting a casing for the rover
The design has been made with Inkscape and can be converted to a DXF
file to be imported by laser cutting software.

It is a box design with screw holes for an Arduino at the bottom and a
Raspberry Pi at the top. The left side of the casing contains holes for
mounting the BattBorg and for wiring.