Test runner: Add arguments to the test.py script
This makes it possible to change some of the previously hardcoded
behavior of the test run script. This includes the file pattern to
retrieve tests from (so we can restrict to certain tests) and the
verbosity of the test result (0 is no dots, 2+ is show all test names).
We can use a different module list to exclude in the import checker.

This commit is mostly a basis for extending the test runner and its
result to process and show more information about test duration, which
can be seen as a test and a benchmark in one.