Bench test result: Add profile output
When the verbosity is 3+, turn on profiling and display the top most
time-taking functions for each test. This means there is a lot of
output, so it works best when restricting the file pattern to certain
tests. Improvements could be made later on to aggregate the results and
get actual benchmarking in the test run.