AStar: Use numpy array distance calculation for close objects
Use the numpy array of nonzero indices in the memory map to calculate
the distance from a current point to all the nonzero points. This
removes the for loop that had an early exit, which takes longer to
calculate a distance because it did not use numpy.

For a low number of nonzero points in the memory map (test_assign), the
performance is about the same, but for a larger number of points, such
as the test_assign_impossible method in the AStar test, the profiler
shows almost a 8x performance improvement for the numpy method.