XBee: Handle exceptions in internal xbee thread
The xbee library creates its own threading object with a running loop
that silences exceptions, except when passing an error callback. Use
this to interrupt the thread manager in case of errors, which can also
happen in the execute method that is also run on that thread.

Add a test to ensure that the receive and error callbacks function
correctly. Use patch contexts so that a test that fail midway still
undoes all its patches.

Ensure we only halt the internal sensor thread during deactivation in
case it is actually alive, otherwise it can cause exceptions in tests
that do not actually start the threading.

The two sources of problems mentioned above caused these tests to hang
when they fail while creating them, without displaying the exceptions.
Even then, a nice trick of our thread manager is that the exceptions are
still logged (and since the test runner hangs, it does not remove them).
This means that debugging it is still possible.