XBee/CC2531 sensor: Clean up some common code
Move out the RSSI broadcast packet handling in the RX process method to
its own method, so that the other packets can be kept in one place in
the physical sensor class.

Also add comments to indicate what each part of the RX process method is
about and which types of packets are involved.