AStar: Add option to allow at bounds
This is useful for grid geometry which has a restricted number of nodes,
but is safe enough to drive on the extremal grid lines. This also fixes
the radius of the region of influence of objects to be similar to how
the closeness worked before. Add a test for the use case of grid
geometry and allow at bounds, where the entire inner part of the grid is
disallowed.