Tests: Replace patcher start/stop with contexts
A unittest patch can be used in multiple ways: as a decorator, as
a context manager in a 'with' statement, as an object during the test or
not at all (i.e. doing the patching ourselves). The first two approaches
have the benefit that if the test fails midway, the "wrapping" decorator
or 'with' context causes the patch to be stopped automatically, thus not
interfering with other tests. Meanwhile, the latter two approaches mean
we have to start and stop manually, which might fail unless it is done
in correctly working setUp/tearDown methods.

Replace the patchers within test methods with context managers, so that
the risk of collateral damage is minimized. Only use patch objects in
setUp/tearDown, and give them unique names. There are cases where the
fourth approach of doing the patching ourselves seems to be necessary,
so those are left as is.

Also remove a stray debugging call in the XBee sensor physical test that
was left over from thread hang tests. The thread manager test superclass
handles the thread checks instead. This commit is mostly a follow-up of
c940a7b411df4f02185748a60373266ae0c58f8e in a larger scope.