Test runner: Suppress mavutil debug print
The pymavlink library has an update to support MAVLink 2.0. For
debugging purposes, the selected dialect's version number is printed
upon an import of the mavutil module. This is bad practice, but we
cannot do much about it right now, so patch the import so that the print
is discarded from the output.