Add pylint configuration
This makes it possible to run pylint on the files or use editor plugins
to run individual reports. Some errors about wacky imports that we need
to do to make the module to behave as a module in-place, as well as the
wacky module name with dashes, have been suppressed. Run it as follows:

pylint mobile-radio-tomography

This command will fill the console with a large volume of lint checks.

Fix one instance in the __init__ file, whether we should fix more of
these or even integrate it into testing, is a matter that needs more
careful changes that do not interfere with existing branches.