USB manager: Handle pySerial 3.1 close exceptions
As of pySerial 3.1.0, blocking reads and writes in serial connections
are more strict about getting closed during the block, causing
exceptions to be thrown more often. This is mostly the case in tests
where a thread is stopped but it has no chance to stop the blocking I/O.
pySerial 3.1 does provide new methods to cancel the blocking read/write
silently, thus making it possible to stop without encountering vague
exceptions about bad file descriptors.

The new pySerial caused Travis to fail intermittently for a short while
again, but this should help in avoiding the intermittent failures.