Vim: Add pylint compiler plugin
This Vim script makes it possible to run pylint from Vim and receive its
linting reports and scores in a quickfix partial window. It is based on
http://www.vim.org/scripts/script.php?script_id=891 with some minor
changes to get all reports in the window instead of just the scoring,
and to make sure the auto-run-on-write only happens in buffer windows
with python file types in them when the compiler is loaded. Otherwise,
when opening another tab/window with a different filetype, saving causes
the default compiler 'make' to be run, which opens the control panel in
a surprise.

Installation instructions and authoring information are in the script.