Tests: Report code coverage
Add a coverage report to the test runner output. This makes it possible
to see which files are not completely covered yet. Fully covered files
are not shown, and we omit other files as well. The report is on by
default for now, but it is skipped when tests fail.