VRMLLoader: Address exception issue
Unfortunately, the PyVRML97 library raises a `NotImplemented` instead of
a `NotImplementedError` in some cases in its `localMatrices` method.
Replace the `isinstance(child, nodetypes.Transforming)` with
a `isinstance(child, basenodes.Transform)`, because the latter always
implements the `localMatrices`. Then we can drop the `try..except`
altogether. We do not need to support any other transforming nodes for
basic VRML support anyway.