Tests: Write test and test base for WiringPi
The core.WiringPi module is somewhat annoying to test in various cases,
because RPi and non-RPi mode are determined in module scope. Therefore,
we need to unload the module every time a test that imports that module
directly or indirectly is run. A base test class provides the utilities
needed for this. Use it in the current tests that have something to do
with WiringPi, namely Environment and Robot_Vehicle. Now the WiringPi
module is fully covered.