WiringPi: Avoid intermittent failures due to module import
Instead of clearing the import of core.WiringPi in the tests, clear the
imports for RPi.GPIO and wiringpi and move those imports to when the
singleton is created. This saves modules from (indirectly) losing their
WiringPi import because we unloaded it in between tests. However, since
RPi.GPIO also raises its error only on the first import, we still need
to clear those imports.