Increase coverage of Arguments and Settings
This depends on the patcher 'with' context managers from an earlier
commit (bae881fbcaf4379be8519351b37b264cc4642562). The changes are
crafted in such a way that there should be minimal conflicts with
changes related to pylint issues. Fix some tests that made a few
incorrect assumptions, and handle all cases in certain methods.