Introduce Import_Manager for dynamic loads
Use the Import_Manager when we want to import a module or class
dynamically based on a string representation, such as a setting
containing the selected class. The Environment tracks an Import_Manager,
passed through to the vehicle for loading the vehicle class. It is also
used for loading the distance sensor, geometry, mission and environment
itself. The Geometry classes are now split into module files, which has
caused quite a number of import changes, unfortunately.

The import manager also provides a (unsafe, but 2.7-standard) way to
unload a module. It thus replaces a bunch of boilerplate code it various
places. More uses, such as dynamically loading the line follower in the
Robot_Vehicle and the XBee class types, follow in other commits.