Pylint: Consider modules from site-packages to be non-standard
Pylint uses an astroid module utility function that guesses whether
a given module is standard (core) or an extension library. The heuristic
uses certain python system configuration information to classify the
module based on the directory it is in. Particularly troublesome are
packages that are installed system-wide, because they may live in either
/usr/lib or /usr/lib64 on 64-bit systems.

Classify any module that has "/site-packages/" in its path as
a non-standard package. Core packages are never installed there, and
this catches all problems that may exist between distributions.

This solves differing pylint warnings about wrong import order,
particularly for packages with compiled libraries.