Tests: Execute pylint on changed files on Travis
Use the TRAVIS_COMMIT_RANGE environment variable as documented on
https://docs.travis-ci.com/user/environment-variables/ and a git call to
determine which files were changed in the commits for which the build is
performed, and run pylint on all python files. Disable long-winded
reports that are of less use when running on a subset of files. Consider
the tests failed if there are any non-info messages, which could be
scaled down to ignore certain message types if necessary.