Pylint: Include files from all commits in branch
Determine the fork point of the branch if it is not the master branch
and we are not performing a pull request build. This causes all files in
the branch to be rerun, ensuring failures in files that are not solved
remain failures. This also prevents problems with rebases that cause
Travis to not know about commits specified in the range.