Arguments: Ensure that choices list is a copy, not a reference
This ensures that the original source of the list of choices is not
changed by other code that adds an empty string to the list when the
option is not required, for example. This would cause weird problems
with __all__ variables containing empty strings.

Test that the original variable is not changed when something is added
to the return value of Arguments.get_choices.