Reconstruction total variation: remove the guess logic
Due to a bug in the code we never actually used the guess. We did
remember the previous image to be used as a guess for the next image,
but in the loop method we would reset it to `None` all the time. Fixing
this issue and using the guess did not produce good results as the image
remains largely the same after approximately three iterations. The
solver apparently thinks it reached a global minimum and stops updating.

Therefore we remove the guess logic entirely since we never used it
anyway. It should help with memory consumption as we do not need to
store the previous frame anymore, as well as runtime performance.