Test runner: Split into multiple files
Move the test runner and test result classes into their own files.
This clears up the test script and makes it possible to add in more
bench-related classes later on.

Add a bit more documentation in the new files, and fix the coverage
`omit` argument so that it excludes the correct files.