Import_Manager: Reload loaded or unloaded modules
This adds functionality to the import manager to be able to reload
modules similar to the builtin reload function, which is mostly useful
to refresh a currently loaded module during runtime (a use case which we
will rarely see in this framework). An additional method allows a reload
of a module that was unloaded in the import manager. This improves the
stability of the unload/reload routine, because modules now remain
tracked instead of possibly losing references. There is still no
guarantee that unloaded modules remain usable, but reloading them can
help in this case. Test the new functionality as well as we can.