Tests: Introduce method coverage
Add a method coverage class with maps test classes and test methods to
corresponding actual classes and methods. This does not work for all
methods, but improvements for that follow later on. The method coverage
report is currently very verbose. The goal is to get more insurance that
every method is tested by "its own" test method (which in some perfectly
reasonable cases can test multiple methods), rather than being tested as
a side-effect of some other test.