Method coverage: Protected method test is optional
Exclude a protected method that is not tested from the method coverage
report. Such a method does not need to be tested directly, since it is
not a part of the public interface. If it is tested by a test, then we
do include it in the totals in the report. We may later determine this
using a setting, so we can see the hidden untested methods. Right now,
there is no need for this though.