Waypoints view: Add compression support
Make it possible to compress and uncompress the waypoints for the
vehicles. This means that subsequent waypoints that can be described as
a range from the previous waypoint to the last waypoint in that sequence
are compressed into one waypoint using the "wait_count" field. The first
waypoint needs special treatment since there is no "previous" waypoint
(or we do not know the home location of the vehicles at that point), but
can still absorb subsequent waypoints that wait at the same location.
The algorithm takes already-compressed waypoints into account, and
calculates the correct new wait count. Also, vehicle wait IDs that
actually differ (not just that one is default and the other is filled)
"break" a range much like a waypoint that just does not fit in it.

Inflating the waypoints is fairly easy in comparison, since it only
needs to check if there is a wait count and generate the range from it.
Uncompressing waypoints also makes the wait IDs explicit if possible.

Add field names to be able retrieve a certain field from waypoint data,
rather than using the index of the column.

Add a method to `Geometry` to check whether two locations are the same,
and test the method. Clean up documentation which omitted some details.